Integer: fully tested

+++ src/Float_Add.cpp
@@ -28,10 +28,10 @@ InfiniteArithmetic::Float InfiniteArithmetic::Float::Add(Float otherOne)
         carry = sum/10;
         result.Array.push_back(digit);
 
-        if(i == thisOne.Array.size() - 1)
+        if(i == (int64_t)thisOne.Array.size() - 1)
             inThis = false;
         
-        if(i == otherOne.Array.size() - 1)
+        if(i == (int64_t)otherOne.Array.size() - 1)
             inOther = false;
     }
